use script with wcp cookie compliance
