--- eng/pipelines/test-deploy.yml
@@ -1,32 +1,0 @@
-resources:
-  repositories:
-  - repository: self
-    ref: $(branch)
-
-pr:
-  paths:
-    include:
-      - eng/pipelines/test-deploy.yml
-
-jobs:
-  - job: build_and_commit
-    displayName: Build and Commit
-    pool:
-      vmImage: 'ubuntu-latest'
-
-    steps:
-    - checkout: self
-
-    - task: UseNode@1
-      inputs:
-        version: '20.x'
-      displayName: 'Install Node.js'
-
-    - task: npmAuthenticate@0
-      inputs:
-        workingFile: ./website/.npmrc
-
-    - script: |
-        npm install && npm run build
-      workingDirectory: website/
-      displayName: 'Build site'
